requirements.txt
allure-pytest
allure-python-commons
allure-python-commons-test
requests
requests-cache
pytest
pytest-pycharm
pytest-xdist
pytest-timeout
pytest-forked
pytest-html
pytest-metadata
pytest-ordering
pytest-reportlog
pytest-rerunfailures
python-dateutil
urllib3
tox
